Added support for different Sink versions
The bracket previously supported PD Buddy Sink versions 0.1 and 0.2,
which were similar enough in size and shape to be able to use the same
bracket.  The current development version of 0.3 is 1 mm narrower, so
the bracket needs a change to support it.  That change has been made,
and a `board` parameter has been added to allow setting the board
revision.  Valid values are `sink_0_1`, `sink_0_2`, and `sink_0_3`.